requirements.txt
alembic==1.5.4
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
integer==1.0.3
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
